Reorganise structure of (sub)modules
Merge `preproc` and `postproc` modules into `utils` and move `standardise.py` and `preprocess.py` to top-level `brainglobe-template-builder`
