+++ brainglobe_template_builder/napari/mask_widget.py
@@ -15,7 +15,7 @@ from qtpy.QtWidgets import (
 
 from brainglobe_template_builder.napari.utils import DirPathWidget
 from brainglobe_template_builder.preproc import create_mask
-from brainglobe_template_builder.preproc.preproc_config import (
+from brainglobe_template_builder.utils.preproc_config import (
     MaskConfig,
     PreprocConfig,
 )
