--- brainglobe_template_builder/preproc/__init__.py
@@ -1,5 +1,0 @@
-from brainglobe_template_builder.preproc.masking import create_mask
-from brainglobe_template_builder.preproc.alignment import (
-    MidplaneAligner,
-    MidplaneEstimator,
-)
