+++ brainglobe_template_builder/utils/annotation_filter.py
