+++ brainglobe_template_builder/utils/brightness.py
