+++ brainglobe_template_builder/utils/cropping.py
