+++ brainglobe_template_builder/utils/masking.py
