+++ brainglobe_template_builder/utils/splitting.py
