+++ examples/blackcap/3_prep_highres.py
@@ -29,7 +29,7 @@ from loguru import logger
 from tqdm import tqdm
 
 from brainglobe_template_builder.io import file_path_with_suffix, load_tiff
-from brainglobe_template_builder.preproc.splitting import (
+from brainglobe_template_builder.utils.splitting import (
     generate_arrays_4template,
     save_array_dict_to_nii,
 )
