+++ examples/molerat/3_molerat_prep_highres.py
@@ -21,8 +21,8 @@ from brainglobe_utils.IO.image.save import save_as_asr_nii
 from loguru import logger
 
 from brainglobe_template_builder.io import file_path_with_suffix, load_tiff
-from brainglobe_template_builder.preproc.masking import create_mask
-from brainglobe_template_builder.preproc.splitting import (
+from brainglobe_template_builder.utils.masking import create_mask
+from brainglobe_template_builder.utils.splitting import (
     get_right_and_left_slices,
     save_array_dict_to_nii,
 )
