+++ examples/tadpole/2_prep_lowres.py
@@ -32,7 +32,7 @@ from tqdm import tqdm
 from brainglobe_template_builder.io import (
     file_path_with_suffix,
 )
-from brainglobe_template_builder.preproc.splitting import (
+from brainglobe_template_builder.utils.splitting import (
     generate_arrays_4template,
     save_array_dict_to_nii,
 )
