+++ tests/test_unit/test_preprocess.py
@@ -11,13 +11,13 @@ from brainglobe_utils.IO.image.save import save_as_asr_nii
 from numpy.testing import assert_raises
 from numpy.typing import NDArray
 
-from brainglobe_template_builder.preproc.preproc_config import PreprocConfig
-from brainglobe_template_builder.preproc.preprocess import (
+from brainglobe_template_builder.preprocess import (
     _create_subject_dir,
     _process_subject,
     _save_niftis,
     preprocess,
 )
+from brainglobe_template_builder.utils.preproc_config import PreprocConfig
 
 
 def make_stack(
